refactor(components): better structure
