+++ front-end/app/components/FlashBlock/styles.css
@@ -1,0 +1,27 @@
+
+.logined-flash {
+  position: fixed;
+  z-index: 10000;
+  top: 30px;
+  left: 50%;
+  transform: translateX(-50%);
+  opacity: 0;
+  width: 200px;
+  padding: 15px;
+  border-radius: 5px;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
+  animation: flash 2s ease-in-out normal;
+}
+
+
+@keyframes flash {
+  from {
+    opacity: 1;
+  }
+
+  to {
+    top: -20px;
+    opacity: 0;
+  }
+
+}
