+++ front-end/app/components/FloatingBlock/styles.css
@@ -1,0 +1,86 @@
+$main-color: #101740;
+$shin-color: #00ffff;
+
+:global {
+  body {
+    margin: 0;
+  }
+}
+
+.floating-block-base {
+  position: fixed;
+  top: 30px;
+  right: 30px;
+  z-index: 1000;
+}
+
+.show-floating-block {
+  composes: floating-block-base;
+  width: 250px;
+  margin: 16px auto;
+  font-size: 30px;
+  text-align: center;
+}
+
+.floating-block {
+  composes: floating-block-base;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
+  transition: box-shadow 0.3s cubic-bezier(.25, .8, .25, 1);
+}
+
+.floating-block:hover {
+  box-shadow: 0 14px 28px rgba(0, 0, 0, 0.25), 0 10px 10px rgba(0, 0, 0, 0.22);
+}
+
+
+.triangle {
+  width: 0;
+  margin-right: auto;
+  margin-left: auto;
+  margin-top: -40px;
+  border: 12px solid transparent;
+  border-bottom-color: $main-color;
+}
+
+.header {
+  background: $main-color;
+  padding: 20px;
+  font-size: 1.4em;
+  font-weight: normal;
+  text-align: center;
+  text-transform: uppercase;
+  color: #fff;
+  margin: 0;
+}
+
+.body {
+  padding: 15px;
+  background: white;
+}
+
+.toggle-display {
+  cursor: pointer;
+}
+
+.toggle-display-word {
+  composes: toggle-display;
+}
+
+.toggle-display-word:hover {
+  text-decoration: underline;
+}
+
+
+.show-floating-block i {
+  transition: all ease-in-out 0.3s;
+  border-radius: 100%;
+}
+
+.show-floating-block i:hover {
+  color: $shin-color;
+  box-shadow: 0px 0px 30px 0px $shin-color;
+}
+
+.toggle-display {
+  cursor: pointer;
+}
