+++ front-end/app/components/StatusBlock/index.js
@@ -1,0 +1,15 @@
+
+import CSSModules from 'react-css-modules'
+
+import styles from './styles.css'
+
+class StatusBlock extends React.Component {
+    
+    render() {
+        return (
+            <div></div>
+            )
+    }
+}
+
+export default CSSModules(StatusBlock, styles)
